deploy/demo_zero_shot/backend/requirements.txt
fastapi
uvicorn
python-multipart
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
onnxruntime>=1.16.0
onnx>=1.14.0
